+++ src/main/java/org/jabref/gui/LibraryTab.java
@@ -136,6 +136,7 @@ public class LibraryTab extends Tab {
         this.getDatabase().registerListener(new SearchListener());
         this.getDatabase().registerListener(new IndexUpdateListener());
         this.getDatabase().registerListener(new EntriesRemovedListener());
+        this.getDatabase().registerListener(new FieldKeywordChangeListener());
 
         // ensure that at each addition of a new entry, the entry is added to the groups interface
         this.bibDatabaseContext.getDatabase().registerListener(new GroupTreeListener());
@@ -825,6 +826,19 @@ public class LibraryTab extends Tab {
             if (preferencesService.getGroupsPreferences().shouldAutoAssignGroup()) {
                 Globals.stateManager.getSelectedGroup(bibDatabaseContext).forEach(
                         selectedGroup -> selectedGroup.addEntriesToGroup(addedEntriesEvent.getBibEntries()));
+            }
+        }
+    }
+
+    /**
+     * Listen whether field of "keywords" changed.
+     */
+    private class FieldKeywordChangeListener {
+
+        @Subscribe
+        public void listen(FieldChangedEvent fieldChangedEvent) {
+            if (fieldChangedEvent.getField().getName().equals("keywords")) {
+                Globals.stateManager.fieldKeywordChangedProperty().set(Optional.of(fieldChangedEvent));
             }
         }
     }